+++ src/modules/TONConfigModule.js
@@ -29,6 +29,8 @@ const DEFAULT_MESSAGE_PROCESSING_TIMEOUT = 40000;
 const DEFAULT_MESSAGE_PROCESSING_GROW_FACTOR = 1.5;
 const DEFAULT_WAIT_FOR_TIMEOUT = 40000;
 
+const DEFAULT_OUT_OF_SYNC_THRESHOLD = 15000;
+
 export class URLParts {
     static parse(url: string): URLParts {
         const protocolSeparatorPos = url.indexOf('://');
@@ -122,6 +124,9 @@ function resolveTimeout(
 }
 
 const defaultServer = 'http://localhost';
+function valueOrDefault(value, defaultValue) {
+    return (value === undefined || value === null) ? defaultValue : value;
+}
 
 export default class TONConfigModule extends TONModule {
     data: TONConfigData;
@@ -143,8 +148,12 @@ export default class TONConfigModule extends TONModule {
     }
 
 
+    outOfSyncThreshold(): number {
+        return valueOrDefault(this.data.outOfSyncThreshold, DEFAULT_OUT_OF_SYNC_THRESHOLD);
+    }
+
     messageRetriesCount(): number {
-        return this.data.messageRetriesCount || DEFAULT_MESSAGE_RETRIES_COUNT;
+        return valueOrDefault(this.data.messageRetriesCount, DEFAULT_MESSAGE_RETRIES_COUNT);
     }
 
     messageExpirationTimeout(retryIndex?: number): number {
@@ -168,7 +177,7 @@ export default class TONConfigModule extends TONModule {
     }
 
     waitForTimeout(): number {
-        return this.data.waitForTimeout || DEFAULT_WAIT_FOR_TIMEOUT;
+        return valueOrDefault(this.data.waitForTimeout, DEFAULT_WAIT_FOR_TIMEOUT);
     }
 
     log(...args: any[]) {